pyproject.toml
[tool.poetry]
name = "vocexcel"
version = "0.4.0"
description = """Another Excel to RDF converter for SKOS vocabs, \
  but one that include profile-based validation of results."""
authors = ["peterphilips <peter.philips@surroundaustralia.com>"]

[tool.poetry.dependencies]
python = "^3.8"
openpyxl = "^3.0.9"
rdflib = "^6.2.0"
pydantic = "^1.9.0"
pyshacl = "^0.18.1"
dateutils = "^0.6.12"
colorama = "^0.4.4"
Pillow = "^9.1.0"
pytest = "^7.2.0"
setuptools = "^62.3.2"
google-api-python-client = "^2.51.0"
pipdeptree = "^2.3.3"
jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "^22.1.0"

[tool.poetry.group.web.dependencies]
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
ruff = "^0.0.277"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vocexcel = "vocexcel.convert:main"

[tool.ruff]
select = [
    "I001",
    "I002"
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]